[fix] The return type of substraitLit.i16() is int32_t, which needs to be cast to int16_t.

+++ cpp-omni/src/substrait/SubstraitToOmniExpr.cpp
@@ -246,7 +246,7 @@ TypedExprPtr SubstraitOmniExprConverter::ToOmniExpr(const ::substrait::Expressio
         case ::substrait::Expression_Literal::LiteralTypeCase::kBoolean:
             return new LiteralExpr(substraitLit.boolean(), BooleanType());
         case ::substrait::Expression_Literal::LiteralTypeCase::kI16:
-            return new LiteralExpr(substraitLit.i16(), ShortType());
+            return new LiteralExpr(static_cast<int16_t>(substraitLit.i16()), ShortType());
         case ::substrait::Expression_Literal::LiteralTypeCase::kI32:
             return new LiteralExpr(substraitLit.i32(), IntType());
         case ::substrait::Expression_Literal::LiteralTypeCase::kI64:
